fix(PPDSC-2031): address comments
