+++ src/audio-player-composable/audio-player-composable.tsx
@@ -17,13 +17,14 @@ export const AudioPlayerComposable = ({
   ariaLandmark,
 }: AudioPlayerComposableProps) => {
   const trackPositionRef = useRef(0);
+
   const audioRef = useRef<HTMLAudioElement>(null);
   const showLoaderTimeoutRef: MutableRefObject<number> = useRef(0);
 
   const [playing, setPlayState] = useState(autoPlay);
   const [loading, setLoading] = useState(true);
   const [duration, setDuration] = useState(0);
-  const [trackPositionArr, setTrackPosition] = useState([0]);
+  const [trackPosition, setTrackPosition] = useState(0);
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const [displayDuration, setDisplayDuration] = useState(0);
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -32,11 +33,11 @@ export const AudioPlayerComposable = ({
   const [buffered, setBuffered] = useState<TimeRanges>();
 
   useEffect(() => {
-    [trackPositionRef.current] = trackPositionArr;
+    trackPositionRef.current = trackPosition;
   });
 
   useEffect(() => {
-    setTrackPosition([0]);
+    setTrackPosition(0);
     setDisplayDuration(0);
   }, [src]);
 
@@ -102,8 +103,8 @@ export const AudioPlayerComposable = ({
 
   const getSeekBarProps = () => ({
     duration,
-    trackPositionArr,
-    onChangeSlider,
+    trackPosition,
+    onChange: onChangeSlider,
     buffered,
   });
