+++ src/audio-player-composable/types.ts
@@ -15,7 +15,7 @@ export interface AudioFunctionDependencies {
   audioRef: React.RefObject<HTMLAudioElement>;
 
   setLoading: React.Dispatch<React.SetStateAction<boolean>>;
-  setTrackPosition: React.Dispatch<React.SetStateAction<number[]>>;
+  setTrackPosition: React.Dispatch<React.SetStateAction<number>>;
   setPlayState: React.Dispatch<React.SetStateAction<boolean>>;
   // setVolume: React.Dispatch<React.SetStateAction<number>>;
   setDuration: React.Dispatch<React.SetStateAction<number>>;
